% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbfgsb3.R
\name{.lbfgsb3cPtr}
\alias{.lbfgsb3cPtr}
\title{Get the lbfgsb3c pointer}
\usage{
.lbfgsb3cPtr()
}
\value{
lbfgsb3c function pointer
}
\description{
Get the lbfgsb3c pointer
}
\examples{

.lbfgsb3cPtr()
}
\author{
Matthew L. Fidler
}
